Make inherent join methods
